+++ src/@types/comment.ts
@@ -1,26 +1,28 @@
+import { CommentLike } from "./commentLike";
 import { Pagination } from "./pagination";
 import { Post } from "./post";
 import { User } from "./user";
 
-export type OnReply = (comment: PostComment) => void;
+export type OnReply = (comment: Comment) => void;
 
 export type NewComment = {
   content: string;
-  parent?: PostComment;
+  parent?: Comment;
 };
 
-export type PostComment = {
+export type Comment = {
   id: string;
-  parent?: PostComment;
+  parent?: Comment;
   user: User;
   post: Post;
-  children?: PostCommentsWithPagination;
+  children?: CommentsWithPagination;
   content: string;
+  isLiked: CommentLike;
   createdAt: string;
   updatedAt: string;
 };
 
-export type PostCommentsWithPagination = {
-  postComments: PostComment[];
+export type CommentsWithPagination = {
+  comments: Comment[];
   pagination: Pagination;
 };