+++ src/@types/post.ts
@@ -1,5 +1,5 @@
+import { CommentsWithPagination } from "./comment";
 import { Pagination } from "./pagination";
-import { PostCommentsWithPagination } from "./postComment";
 import { PostLike, PostLikesWithPagination } from "./postLike";
 import { PostSave } from "./postSave";
 import { User } from "./user";
@@ -12,7 +12,7 @@ export type Post = {
   isLiked?: PostLike;
   isSaved?: PostSave;
   likes?: PostLikesWithPagination;
-  comments?: PostCommentsWithPagination;
+  comments?: CommentsWithPagination;
   createdAt: string;
   updatedAt: string;
 };