+++ src/components/ReplyComment/@types/index.ts
@@ -1,6 +1,0 @@
-import { OnReply, PostComment } from "@t/postComment";
-
-export type ReplyCommentProps = {
-  comment: PostComment;
-  onReply: OnReply;
-};