+++ src/features/Comment/hooks/index.ts
@@ -1,0 +1,5 @@
+import { useComment } from "./useComment";
+import { useCommentLike } from "./useCommentLike";
+import { useComments } from "./useComments";
+
+export { useComment, useComments, useCommentLike };