+++ src/features/Post/hooks/index.ts
@@ -1,5 +1,6 @@
+import { useAddComment } from "./useAddComment";
 import { usePost } from "./usePost";
 import { usePostLike } from "./usePostLike";
 import { usePostSave } from "./usePostSave";
 
-export { usePost, usePostLike, usePostSave };
+export { usePost, usePostLike, usePostSave, useAddComment };