+++ src/graphql/@types/comment.ts
@@ -1,0 +1,9 @@
+import { Comment, CommentsWithPagination } from "@t/comment";
+
+export type AddCommentMutation = {
+  addComment: Comment;
+};
+
+export type GetCommentsQuery = {
+  getComments: CommentsWithPagination;
+};