+++ src/graphql/@types/commentLike.ts
@@ -1,0 +1,9 @@
+import { CommentLike } from "@t/commentLike";
+
+export type LikeCommentMutation = {
+  likeComment: CommentLike;
+};
+
+export type UnlikeCommentMutation = {
+  unlikeComment: CommentLike;
+};