+++ src/graphql/post.ts
@@ -25,20 +25,26 @@ export const GET_POST_BY_URL = gql`
         }
       }
       comments {
-        postComments {
+        comments {
           id
           user {
             id
             username
             fullname
           }
+          isLiked {
+            id
+          }
           children {
-            postComments {
+            comments {
               id
               user {
                 id
                 username
                 fullname
+              }
+              isLiked {
+                id
               }
               content
               updatedAt