+++ src/graphql/uploadFile.ts
@@ -1,0 +1,9 @@
+import { gql } from "@apollo/client";
+
+export const SINGLE_UPLOAD = gql`
+  mutation singleUpload($file: Upload!) {
+    singleUpload(file: $file) {
+      name
+    }
+  }
+`;