+++ CHANGELOG.md
@@ -15,7 +15,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 * add commentCounter for PostCard feature ([019855c](https://github.com/favecode/plog-fe/commit/019855c5cc7e4a3e83c08fcdcced5a0d74d580cc))
 * add dark mode for storybook ([8fb6f34](https://github.com/favecode/plog-fe/commit/8fb6f349d345ce036ae903056402ecb5267153f3))
-* add postComment queries ([669fce8](https://github.com/favecode/plog-fe/commit/669fce8f0cd8095c0d2a9b4a7ba7ee448b80670f))
+* add comment queries ([669fce8](https://github.com/favecode/plog-fe/commit/669fce8f0cd8095c0d2a9b4a7ba7ee448b80670f))
 * implement Comment component ([8107ce2](https://github.com/favecode/plog-fe/commit/8107ce2cb166b1f766d9fb0151cc42ef5fc7e452))
 * implement Post page ([50bd65e](https://github.com/favecode/plog-fe/commit/50bd65e245517c21b1075a579d9a5f9baac50b87))